Renamed minplus to algebra to make room for maxplus, transforms etc.
Also renamed dnc to disco to make the origin of the implementation clear
/ contrasting it to mpa rtc
