+++ src/main/java/de/uni_kl/cs/discodnc/Calculator.java
@@ -28,8 +28,8 @@
 
 package de.uni_kl.cs.discodnc;
 
+import de.uni_kl.cs.discodnc.algebra.MinPlus;
 import de.uni_kl.cs.discodnc.curves.Curve;
-import de.uni_kl.cs.discodnc.minplus.MinPlus;
 import de.uni_kl.cs.discodnc.numbers.NumBackend;
 
 public final class Calculator {
