+++ src/main/java/de/uni_kl/cs/discodnc/CurveBackend.java
@@ -27,9 +27,9 @@
 
 package de.uni_kl.cs.discodnc;
 
+import de.uni_kl.cs.discodnc.algebra.MinPlus;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.LinearSegment;
-import de.uni_kl.cs.discodnc.minplus.MinPlus;
 
 public interface CurveBackend {
 	MinPlus getMinPlus();
