+++ src/main/java/de/uni_kl/cs/discodnc/curves/disco/affine/ArrivalCurve_DNC_Affine.java
@@ -27,44 +27,52 @@
  *
  */
 
-package de.uni_kl.cs.discodnc.curves.dnc.pwaffine;
+package de.uni_kl.cs.discodnc.curves.disco.affine;
 
 import de.uni_kl.cs.discodnc.Calculator;
+import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
-import de.uni_kl.cs.discodnc.curves.ServiceCurve;
 
-public class ServiceCurve_DNC_PwAffine extends Curve_DNC_PwAffine implements ServiceCurve {
+public class ArrivalCurve_DNC_Affine extends Curve_DNC_Affine implements ArrivalCurve {
     // --------------------------------------------------------------------------------------------------------------
     // Constructors
     // --------------------------------------------------------------------------------------------------------------
-    public ServiceCurve_DNC_PwAffine() {
+    public ArrivalCurve_DNC_Affine() {
         super();
     }
 
-    public ServiceCurve_DNC_PwAffine(int segment_count) {
+    public ArrivalCurve_DNC_Affine(int segment_count) {
         super(segment_count);
     }
 
-    public ServiceCurve_DNC_PwAffine(Curve curve) {
-        copy(curve);
+    public ArrivalCurve_DNC_Affine(Curve curve) {
+        super(curve);
+        forceThroughOrigin();
 
-        // Too strong requirement: !isConvex()
-        if (Calculator.getInstance().exec_service_curve_checks() && !isWideSenseIncreasing()) {
-            throw new RuntimeException("Service curves can only be created from wide-sense increasing functions.");
+        if (Calculator.getInstance().exec_arrival_curve_checks() && !isWideSenseIncreasing()) { // too strong
+            // requirement:
+            // !isConcave()
+            System.out.println(toString());
+            throw new RuntimeException("Arrival curves can only be created from wide-sense increasing functions.");
         }
     }
 
-    public ServiceCurve_DNC_PwAffine(String service_curve_str) throws Exception {
-        if (service_curve_str == null || service_curve_str.isEmpty() || service_curve_str.length() < 9) {
-        	// Smallest possible string: {(0,0),0}
+    public ArrivalCurve_DNC_Affine(String arrival_curve_str) throws Exception {
+        if (arrival_curve_str == null || arrival_curve_str.isEmpty() || arrival_curve_str.length() < 9) { // Smallest
+            // possible
+            // string:
+            // {(0,0),0}
             throw new RuntimeException("Invalid string representation of a service curve.");
         }
 
-        initializeCurve(service_curve_str);
+        initializeCurve(arrival_curve_str);
+        forceThroughOrigin();
 
-        // Too strong requirement: !isConvex()
-        if (Calculator.getInstance().exec_service_curve_checks() && !isWideSenseIncreasing()) {
-            throw new RuntimeException("Service curves can only be created from wide-sense increasing functions.");
+        if (Calculator.getInstance().exec_arrival_curve_checks() && !isWideSenseIncreasing()) { // too strong
+            // requirement:
+            // !isConcave()
+            System.out.println(toString());
+            throw new RuntimeException("Arrival curves can only be created from wide-sense increasing functions.");
         }
     }
 
@@ -72,20 +80,20 @@ public class ServiceCurve_DNC_PwAffine extends Curve_DNC_PwAffine implements Ser
     // Interface Implementations
     // --------------------------------------------------------------------------------------------------------------
     @Override
-    public ServiceCurve_DNC_PwAffine copy() {
-        ServiceCurve_DNC_PwAffine sc_copy = new ServiceCurve_DNC_PwAffine();
-        sc_copy.copy(this);
-        return sc_copy;
+    public ArrivalCurve_DNC_Affine copy() {
+        ArrivalCurve_DNC_Affine ac_copy = new ArrivalCurve_DNC_Affine();
+        ac_copy.copy(this);
+        return ac_copy;
     }
 
     @Override
     public boolean equals(Object obj) {
-        return (obj instanceof ServiceCurve_DNC_PwAffine) && super.equals(obj);
+        return (obj instanceof ArrivalCurve_DNC_Affine) && super.equals(obj);
     }
 
     @Override
     public int hashCode() {
-        return "SC".hashCode() * super.hashCode();
+        return "AC".hashCode() * super.hashCode();
     }
 
     /**
@@ -95,6 +103,6 @@ public class ServiceCurve_DNC_PwAffine extends Curve_DNC_PwAffine implements Ser
      */
     @Override
     public String toString() {
-        return "SC" + super.toString();
+        return "AC" + super.toString();
     }
 }
