+++ src/main/java/de/uni_kl/cs/discodnc/curves/disco/pwaffine/Curve_DNC_PwAffine.java
@@ -27,13 +27,13 @@
  *
  */
 
-package de.uni_kl.cs.discodnc.curves.dnc.affine;
+package de.uni_kl.cs.discodnc.curves.disco.pwaffine;
 
 import de.uni_kl.cs.discodnc.Calculator;
 import de.uni_kl.cs.discodnc.curves.Curve;
-import de.uni_kl.cs.discodnc.curves.Curve_Affine;
+import de.uni_kl.cs.discodnc.curves.Curve_PwAffine;
 import de.uni_kl.cs.discodnc.curves.LinearSegment;
-import de.uni_kl.cs.discodnc.curves.dnc.LinearSegment_DNC;
+import de.uni_kl.cs.discodnc.curves.disco.LinearSegment_DNC;
 import de.uni_kl.cs.discodnc.numbers.Num;
 
 import java.util.ArrayList;
@@ -53,8 +53,8 @@ import java.util.List;
  * All arithmetic operations on a curve return a new instance of class
  * <code>Curve</code>.<br>
  */
-public class Curve_DNC_Affine implements Curve_Affine {
-	private static Curve_DNC_Affine instance = new Curve_DNC_Affine();
+public class Curve_DNC_PwAffine implements Curve_PwAffine {
+	private static Curve_DNC_PwAffine instance = new Curve_DNC_PwAffine();
 
 	protected LinearSegment_DNC[] segments;
 
@@ -62,20 +62,20 @@ public class Curve_DNC_Affine implements Curve_Affine {
 
 	protected boolean is_rate_latency = false;
 	protected boolean has_rate_latency_meta_info = false;
-	protected List<Curve_DNC_Affine> rate_latencies = new LinkedList<Curve_DNC_Affine>();
+	protected List<Curve_DNC_PwAffine> rate_latencies = new LinkedList<Curve_DNC_PwAffine>();
 
 	protected boolean is_token_bucket = false;
 	protected boolean has_token_bucket_meta_info = false;
-	protected List<Curve_DNC_Affine> token_buckets = new LinkedList<Curve_DNC_Affine>();
+	protected List<Curve_DNC_PwAffine> token_buckets = new LinkedList<Curve_DNC_PwAffine>();
 
 	/**
 	 * Creates a <code>CurveDNC</code> instance with a single segment on the x-axis.
 	 */
-	protected Curve_DNC_Affine() {
+	protected Curve_DNC_PwAffine() {
 		createNewCurve(1, false);
 	}
 
-	protected Curve_DNC_Affine(Curve curve) {
+	protected Curve_DNC_PwAffine(Curve curve) {
 		copy(curve);
 	}
 
@@ -86,15 +86,11 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	 * @param segment_count
 	 *            the number of segments
 	 */
-	protected Curve_DNC_Affine(int segment_count) {
-		// // TODO requires fix to Curve's boundAtXAxis.
-//		if (segment_count < 0 || segment_count > 2) {
-//			throw new IndexOutOfBoundsException("Affine curves can have at most two segments (given count was " + segment_count + ")!");
-//		}
+	protected Curve_DNC_PwAffine(int segment_count) {
 		createNewCurve(segment_count, false);
 	}
 
-	public static Curve_DNC_Affine getFactory() {
+	public static Curve_DNC_PwAffine getFactory() {
 		return instance;
 	}
 
@@ -120,8 +116,8 @@ public class Curve_DNC_Affine implements Curve_Affine {
 		this.has_rate_latency_meta_info = has_rate_latency_meta_info;
 	}
 
-	public List<Curve_Affine> getRL_Components() {
-		List<Curve_Affine> tmp = new LinkedList<>();
+	public List<Curve_PwAffine> getRL_Components() {
+		List<Curve_PwAffine> tmp = new LinkedList<>();
 		if (this.is_rate_latency) {
 			tmp.add(this.copy());
 		} else {
@@ -133,9 +129,9 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	}
 
 	public void setRL_Components(List<Curve> rate_latencies) {
-		List<Curve_DNC_Affine> tmp = new LinkedList<>();
+		List<Curve_DNC_PwAffine> tmp = new LinkedList<>();
 		for (int i = 0; i < rate_latencies.size(); i++) {
-			tmp.add((Curve_DNC_Affine) rate_latencies.get(i));
+			tmp.add((Curve_DNC_PwAffine) rate_latencies.get(i));
 		}
 		this.rate_latencies = tmp;
 	}
@@ -148,8 +144,8 @@ public class Curve_DNC_Affine implements Curve_Affine {
 		this.has_token_bucket_meta_info = has_token_bucket_meta_info;
 	}
 
-	public List<Curve_Affine> getTB_Components() {
-		List<Curve_Affine> tmp = new LinkedList<>();
+	public List<Curve_PwAffine> getTB_Components() {
+		List<Curve_PwAffine> tmp = new LinkedList<>();
 		for (int i = 0; i < token_buckets.size(); i++) {
 			tmp.add(token_buckets.get(i));
 		}
@@ -157,9 +153,9 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	}
 
 	public void setTB_Components(List<Curve> token_buckets) {
-		List<Curve_DNC_Affine> tmp = new LinkedList<>();
+		List<Curve_DNC_PwAffine> tmp = new LinkedList<>();
 		for (int i = 0; i < token_buckets.size(); i++) {
-			tmp.add((Curve_DNC_Affine) token_buckets.get(i));
+			tmp.add((Curve_DNC_PwAffine) token_buckets.get(i));
 		}
 		this.token_buckets = tmp;
 	}
@@ -237,11 +233,11 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	private void clearMetaInfo() {
 		has_token_bucket_meta_info = false;
 		is_token_bucket = false;
-		token_buckets = new LinkedList<Curve_DNC_Affine>();
+		token_buckets = new LinkedList<Curve_DNC_PwAffine>();
 
 		has_rate_latency_meta_info = false;
 		is_rate_latency = false;
-		rate_latencies = new LinkedList<Curve_DNC_Affine>();
+		rate_latencies = new LinkedList<Curve_DNC_PwAffine>();
 	}
 
 	/**
@@ -250,8 +246,8 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	 * @return a copy of this instance.
 	 */
 	@Override
-	public Curve_DNC_Affine copy() {
-		Curve_DNC_Affine c_copy = new Curve_DNC_Affine();
+	public Curve_DNC_PwAffine copy() {
+		Curve_DNC_PwAffine c_copy = new Curve_DNC_PwAffine();
 		c_copy.copy(this);
 		return c_copy;
 	}
@@ -260,20 +256,20 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	public void copy(Curve curve) {
 		LinearSegment_DNC[] segments = new LinearSegment_DNC[curve.getSegmentCount()];
 
-		if (curve instanceof Curve_DNC_Affine) {
+		if (curve instanceof Curve_DNC_PwAffine) {
 			for (int i = 0; i < segments.length; i++) {
-				segments[i] = ((Curve_DNC_Affine) curve).getSegment(i).copy();
+				segments[i] = ((Curve_DNC_PwAffine) curve).getSegment(i).copy();
 			}
 
-			this.has_rate_latency_meta_info = ((Curve_DNC_Affine) curve).has_rate_latency_meta_info;
-			this.rate_latencies = ((Curve_DNC_Affine) curve).rate_latencies;
+			this.has_rate_latency_meta_info = ((Curve_DNC_PwAffine) curve).has_rate_latency_meta_info;
+			this.rate_latencies = ((Curve_DNC_PwAffine) curve).rate_latencies;
 
-			this.has_token_bucket_meta_info = ((Curve_DNC_Affine) curve).has_token_bucket_meta_info;
-			this.token_buckets = ((Curve_DNC_Affine) curve).token_buckets;
+			this.has_token_bucket_meta_info = ((Curve_DNC_PwAffine) curve).has_token_bucket_meta_info;
+			this.token_buckets = ((Curve_DNC_PwAffine) curve).token_buckets;
 
-			this.is_delayed_infinite_burst = ((Curve_Affine) curve).isDelayedInfiniteBurst();
-			this.is_rate_latency = ((Curve_Affine) curve).isRateLatency();
-			this.is_token_bucket = ((Curve_Affine) curve).isTokenBucket();
+			this.is_delayed_infinite_burst = ((Curve_PwAffine) curve).isDelayedInfiniteBurst();
+			this.is_rate_latency = ((Curve_PwAffine) curve).isRateLatency();
+			this.is_token_bucket = ((Curve_PwAffine) curve).isTokenBucket();
 		} else {
 			for (int i = 0; i < curve.getSegmentCount(); i++) {
 				segments[i] = new LinearSegment_DNC(curve.getSegment(i));
@@ -543,7 +539,8 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	 */
 	public boolean isConvexIn(Num a, Num b) {
 		Num last_gradient = Num.getFactory(Calculator.getInstance().getNumBackend()).getNegativeInfinity(); // No need to create an object as this
-		// value is only set for initial comparison in the loop.
+		// value is only set for initial
+		// comparison in the loop.
 
 		int i_start = getSegmentDefining(a);
 		int i_end = getSegmentDefining(b);
@@ -589,7 +586,8 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	 */
 	public boolean isConcaveIn(Num a, Num b) {
 		Num last_gradient = Num.getFactory(Calculator.getInstance().getNumBackend()).getPositiveInfinity(); // No need to create an object as this
-		// value is only set for initial comparison in the loop.
+		// value is only set for initial
+		// comparison in the loop.
 
 		int i_start = getSegmentDefining(a);
 		int i_end = getSegmentDefining(b);
@@ -621,7 +619,8 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	 */
 	public boolean isAlmostConcave() {
 		Num last_gradient = Num.getFactory(Calculator.getInstance().getNumBackend()).getPositiveInfinity(); // No need to create an object as this
-		// value is only set for initial comparison in the loop.
+		// value is only set for initial
+		// comparison in the loop.
 
 		for (int i = 0; i < segments.length; i++) {
 			// Skip the horizontal part at the beginning
@@ -647,12 +646,12 @@ public class Curve_DNC_Affine implements Curve_Affine {
 
 	@Override
 	public boolean equals(Object obj) {
-		if (obj == null || !(obj instanceof Curve_DNC_Affine)) {
+		if (obj == null || !(obj instanceof Curve_DNC_PwAffine)) {
 			return false;
 		}
 
-		Curve_DNC_Affine this_cpy = this.copy();
-		Curve_DNC_Affine other_cpy = ((Curve_DNC_Affine) obj).copy();
+		Curve_DNC_PwAffine this_cpy = this.copy();
+		Curve_DNC_PwAffine other_cpy = ((Curve_DNC_PwAffine) obj).copy();
 
 		Curve.beautify(this_cpy);
 		Curve.beautify(other_cpy);
@@ -923,7 +922,7 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	 *            the number of the rate latency curve
 	 * @return the rate latency curve
 	 */
-	public Curve_DNC_Affine getRL_Component(int i) {
+	public Curve_DNC_PwAffine getRL_Component(int i) {
 		decomposeIntoRateLatencies();
 		return rate_latencies.get(i);
 	}
@@ -940,14 +939,14 @@ public class Curve_DNC_Affine implements Curve_Affine {
 
 		if (Calculator.getInstance().exec_service_curve_checks() && !this.isConvex()) {
 			if (this.equals(this.createZeroDelayInfiniteBurst())) {
-				rate_latencies = new ArrayList<Curve_DNC_Affine>();
+				rate_latencies = new ArrayList<Curve_DNC_PwAffine>();
 				rate_latencies.add(this.createRateLatency(Num.getFactory(Calculator.getInstance().getNumBackend()).createPositiveInfinity(),
 						Num.getFactory(Calculator.getInstance().getNumBackend()).createZero()));
 			} else {
 				throw new RuntimeException("Can only decompose convex service curves into rate latency curves.");
 			}
 		} else {
-			rate_latencies = new ArrayList<Curve_DNC_Affine>();
+			rate_latencies = new ArrayList<Curve_DNC_PwAffine>();
 			for (int i = 0; i < segments.length; i++) {
 				if (segments[i].getY().eq(0.0) && segments[i].getGrad().eq(0.0)) {
 					continue;
@@ -989,7 +988,7 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	 *            the number of the token bucket
 	 * @return the token bucket
 	 */
-	public Curve_DNC_Affine getTB_Component(int i) {
+	public Curve_DNC_PwAffine getTB_Component(int i) {
 		decomposeIntoTokenBuckets();
 		return token_buckets.get(i);
 	}
@@ -1008,7 +1007,7 @@ public class Curve_DNC_Affine implements Curve_Affine {
 			throw new RuntimeException("Can only decompose concave arrival curves into token buckets.");
 		}
 
-		token_buckets = new ArrayList<Curve_DNC_Affine>();
+		token_buckets = new ArrayList<Curve_DNC_PwAffine>();
 		for (int i = 0; i < segments.length; i++) {
 			if (isDiscontinuity(i)) {
 				continue;
@@ -1035,16 +1034,8 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	// ------------------------------------------------------------
 	// DiscoDNC compliance
 	// ------------------------------------------------------------
-	public Curve_DNC_Affine createCurve(List<LinearSegment> segments) {
-		// TODO Assume there are more than two segments, defining either a valid mTB or mRL.
-		// The first segment passes through the origin, the second defines best known lateny or burstiness.
-		// The last segment defines least arrival rate or largest service rate, respectively.
-		// Choosing the second segment for an affine curve from the given ones is thus finding a reasonable tradeoff.
-		// We should trade accuracy for certainty to not cause false resource exhaustion assumptions.
-		// I.e., the second segment of our affine curve should be defined by the last one of the givens segments by
-		// extending it to the front such that it either intersects with, i.e.,, then starts at,
-		// the x-axis (defining a new service curve latency) or the y-axis (defining a new arrival curve burstiness).
-		Curve_DNC_Affine c_dnc = new Curve_DNC_Affine(segments.size());
+	public Curve_DNC_PwAffine createCurve(List<LinearSegment> segments) {
+		Curve_DNC_PwAffine c_dnc = new Curve_DNC_PwAffine(segments.size());
 		for (int i = 0; i < segments.size(); i++) {
 			c_dnc.setSegment(i, segments.get(i));
 		}
@@ -1052,11 +1043,11 @@ public class Curve_DNC_Affine implements Curve_Affine {
 		return c_dnc;
 	}
 
-	public Curve_DNC_Affine createZeroCurve() {
-		return new Curve_DNC_Affine(); // CurveDNC constructor's default behavior
+	public Curve_DNC_PwAffine createZeroCurve() {
+		return new Curve_DNC_PwAffine(); // CurveDNC constructor's default behavior
 	}
 
-	public Curve_DNC_Affine createHorizontal(double y) {
+	public Curve_DNC_PwAffine createHorizontal(double y) {
 		return createHorizontal(Num.getFactory(Calculator.getInstance().getNumBackend()).create(y));
 	}
 
@@ -1067,8 +1058,8 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	 *            the y-intercept of the curve
 	 * @return a <code>Curve</code> instance
 	 */
-	public Curve_DNC_Affine createHorizontal(Num y) {
-		Curve_DNC_Affine c_dnc = new Curve_DNC_Affine();
+	public Curve_DNC_PwAffine createHorizontal(Num y) {
+		Curve_DNC_PwAffine c_dnc = new Curve_DNC_PwAffine();
 		makeHorizontal(c_dnc, y);
 		return c_dnc;
 	}
@@ -1080,24 +1071,24 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	// ------------------------------------------------------------
 	// DiscoDNC compliance
 	// ------------------------------------------------------------
-	public ServiceCurve_DNC_Affine createServiceCurve() {
-		return new ServiceCurve_DNC_Affine();
+	public ServiceCurve_DNC_PwAffine createServiceCurve() {
+		return new ServiceCurve_DNC_PwAffine();
 	}
 
-	public ServiceCurve_DNC_Affine createServiceCurve(int segment_count) {
-		return new ServiceCurve_DNC_Affine(segment_count);
+	public ServiceCurve_DNC_PwAffine createServiceCurve(int segment_count) {
+		return new ServiceCurve_DNC_PwAffine(segment_count);
 	}
 
-	public ServiceCurve_DNC_Affine createServiceCurve(String service_curve_str) throws Exception {
-		return new ServiceCurve_DNC_Affine(service_curve_str);
+	public ServiceCurve_DNC_PwAffine createServiceCurve(String service_curve_str) throws Exception {
+		return new ServiceCurve_DNC_PwAffine(service_curve_str);
 	}
 
-	public ServiceCurve_DNC_Affine createServiceCurve(Curve curve) {
-		return new ServiceCurve_DNC_Affine(curve);
+	public ServiceCurve_DNC_PwAffine createServiceCurve(Curve curve) {
+		return new ServiceCurve_DNC_PwAffine(curve);
 	}
 
-	public ServiceCurve_DNC_Affine createZeroService() {
-		return new ServiceCurve_DNC_Affine(); // ServiceCurveDNC constructor's default behavior
+	public ServiceCurve_DNC_PwAffine createZeroService() {
+		return new ServiceCurve_DNC_PwAffine(); // ServiceCurveDNC constructor's default behavior
 	}
 
 	/**
@@ -1105,26 +1096,26 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	 *
 	 * @return a <code>ServiceCurve</code> instance
 	 */
-	public ServiceCurve_DNC_Affine createZeroDelayInfiniteBurst() {
+	public ServiceCurve_DNC_PwAffine createZeroDelayInfiniteBurst() {
 		return createDelayedInfiniteBurst(Num.getFactory(Calculator.getInstance().getNumBackend()).createZero());
 	}
 
-	public ServiceCurve_DNC_Affine createDelayedInfiniteBurst(double delay) {
+	public ServiceCurve_DNC_PwAffine createDelayedInfiniteBurst(double delay) {
 		return createDelayedInfiniteBurst(Num.getFactory(Calculator.getInstance().getNumBackend()).create(delay));
 	}
 
-	public ServiceCurve_DNC_Affine createDelayedInfiniteBurst(Num delay) {
-		ServiceCurve_DNC_Affine sc_dnc = new ServiceCurve_DNC_Affine();
+	public ServiceCurve_DNC_PwAffine createDelayedInfiniteBurst(Num delay) {
+		ServiceCurve_DNC_PwAffine sc_dnc = new ServiceCurve_DNC_PwAffine();
 		makeDelayedInfiniteBurst(sc_dnc, delay);
 		return sc_dnc;
 	}
 
-	public ServiceCurve_DNC_Affine createRateLatency(double rate, double latency) {
+	public ServiceCurve_DNC_PwAffine createRateLatency(double rate, double latency) {
 		return createRateLatency(Num.getFactory(Calculator.getInstance().getNumBackend()).create(rate), Num.getFactory(Calculator.getInstance().getNumBackend()).create(latency));
 	}
 
-	public ServiceCurve_DNC_Affine createRateLatency(Num rate, Num latency) {
-		ServiceCurve_DNC_Affine sc_dnc = new ServiceCurve_DNC_Affine();
+	public ServiceCurve_DNC_PwAffine createRateLatency(Num rate, Num latency) {
+		ServiceCurve_DNC_PwAffine sc_dnc = new ServiceCurve_DNC_PwAffine();
 		makeRateLatency(sc_dnc, rate, latency);
 		return sc_dnc;
 	}
@@ -1136,52 +1127,52 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	// ------------------------------------------------------------
 	// DiscoDNC compliance
 	// ------------------------------------------------------------
-	public ArrivalCurve_DNC_Affine createArrivalCurve() {
-		return new ArrivalCurve_DNC_Affine();
+	public ArrivalCurve_DNC_PwAffine createArrivalCurve() {
+		return new ArrivalCurve_DNC_PwAffine();
 	}
 
-	public ArrivalCurve_DNC_Affine createArrivalCurve(int segment_count) {
-		return new ArrivalCurve_DNC_Affine(segment_count);
+	public ArrivalCurve_DNC_PwAffine createArrivalCurve(int segment_count) {
+		return new ArrivalCurve_DNC_PwAffine(segment_count);
 	}
 
-	public ArrivalCurve_DNC_Affine createArrivalCurve(String arrival_curve_str) throws Exception {
-		return new ArrivalCurve_DNC_Affine(arrival_curve_str);
+	public ArrivalCurve_DNC_PwAffine createArrivalCurve(String arrival_curve_str) throws Exception {
+		return new ArrivalCurve_DNC_PwAffine(arrival_curve_str);
 	}
 
-	public ArrivalCurve_DNC_Affine createArrivalCurve(Curve curve) {
-		return new ArrivalCurve_DNC_Affine(curve);
+	public ArrivalCurve_DNC_PwAffine createArrivalCurve(Curve curve) {
+		return new ArrivalCurve_DNC_PwAffine(curve);
 	}
 
-	public ArrivalCurve_DNC_Affine createArrivalCurve(Curve curve, boolean remove_latency) {
+	public ArrivalCurve_DNC_PwAffine createArrivalCurve(Curve curve, boolean remove_latency) {
 		return createArrivalCurve(Curve.removeLatency(curve));
 	}
 
-	public ArrivalCurve_DNC_Affine createZeroArrivals() {
-		return new ArrivalCurve_DNC_Affine(); // ArrivalCurveDNC constructor's default behavior
+	public ArrivalCurve_DNC_PwAffine createZeroArrivals() {
+		return new ArrivalCurve_DNC_PwAffine(); // ArrivalCurveDNC constructor's default behavior
 	}
 
-	public ArrivalCurve_DNC_Affine createInfiniteArrivals() {
-		ArrivalCurve_DNC_Affine ac_dnc = new ArrivalCurve_DNC_Affine(); 
+	public ArrivalCurve_DNC_PwAffine createInfiniteArrivals() {
+		ArrivalCurve_DNC_PwAffine ac_dnc = new ArrivalCurve_DNC_PwAffine(); 
 		makeDelayedInfiniteBurst(ac_dnc, Num.getFactory(Calculator.getInstance().getNumBackend()).createZero());
 		return ac_dnc;
 	}
 
-	public ArrivalCurve_DNC_Affine createPeakArrivalRate(double rate) {
+	public ArrivalCurve_DNC_PwAffine createPeakArrivalRate(double rate) {
 		return createPeakArrivalRate(Num.getFactory(Calculator.getInstance().getNumBackend()).create(rate));
 	}
 
-	public ArrivalCurve_DNC_Affine createPeakArrivalRate(Num rate) {
-		ArrivalCurve_DNC_Affine ac_dnc = new ArrivalCurve_DNC_Affine();
+	public ArrivalCurve_DNC_PwAffine createPeakArrivalRate(Num rate) {
+		ArrivalCurve_DNC_PwAffine ac_dnc = new ArrivalCurve_DNC_PwAffine();
 		makePeakRate(ac_dnc, rate);
 		return ac_dnc;
 	}
 
-	public ArrivalCurve_DNC_Affine createTokenBucket(double rate, double burst) {
+	public ArrivalCurve_DNC_PwAffine createTokenBucket(double rate, double burst) {
 		return createTokenBucket(Num.getFactory(Calculator.getInstance().getNumBackend()).create(rate), Num.getFactory(Calculator.getInstance().getNumBackend()).create(burst));
 	}
 
-	public ArrivalCurve_DNC_Affine createTokenBucket(Num rate, Num burst) {
-		ArrivalCurve_DNC_Affine ac_dnc = new ArrivalCurve_DNC_Affine();
+	public ArrivalCurve_DNC_PwAffine createTokenBucket(Num rate, Num burst) {
+		ArrivalCurve_DNC_PwAffine ac_dnc = new ArrivalCurve_DNC_PwAffine();
 		makeTokenBucket(ac_dnc, rate, burst);
 		return ac_dnc;
 	}
@@ -1193,46 +1184,46 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	// ------------------------------------------------------------
 	// DiscoDNC compliance
 	// ------------------------------------------------------------
-	public MaxServiceCurve_DNC_Affine createMaxServiceCurve() {
-		return new MaxServiceCurve_DNC_Affine();
+	public MaxServiceCurve_DNC_PwAffine createMaxServiceCurve() {
+		return new MaxServiceCurve_DNC_PwAffine();
 	}
 
-	public MaxServiceCurve_DNC_Affine createMaxServiceCurve(int segment_count) {
-		return new MaxServiceCurve_DNC_Affine(segment_count);
+	public MaxServiceCurve_DNC_PwAffine createMaxServiceCurve(int segment_count) {
+		return new MaxServiceCurve_DNC_PwAffine(segment_count);
 	}
 
-	public MaxServiceCurve_DNC_Affine createMaxServiceCurve(String max_service_curve_str) throws Exception {
-		return new MaxServiceCurve_DNC_Affine(max_service_curve_str);
+	public MaxServiceCurve_DNC_PwAffine createMaxServiceCurve(String max_service_curve_str) throws Exception {
+		return new MaxServiceCurve_DNC_PwAffine(max_service_curve_str);
 	}
 
-	public MaxServiceCurve_DNC_Affine createMaxServiceCurve(Curve curve) {
-		return new MaxServiceCurve_DNC_Affine(curve);
+	public MaxServiceCurve_DNC_PwAffine createMaxServiceCurve(Curve curve) {
+		return new MaxServiceCurve_DNC_PwAffine(curve);
 	}
 
-	public MaxServiceCurve_DNC_Affine createInfiniteMaxService() {
+	public MaxServiceCurve_DNC_PwAffine createInfiniteMaxService() {
 		return createDelayedInfiniteBurstMSC(Num.getFactory(Calculator.getInstance().getNumBackend()).createZero());
 	}
 
-	public MaxServiceCurve_DNC_Affine createZeroDelayInfiniteBurstMSC() {
+	public MaxServiceCurve_DNC_PwAffine createZeroDelayInfiniteBurstMSC() {
 		return createDelayedInfiniteBurstMSC(Num.getFactory(Calculator.getInstance().getNumBackend()).createZero());
 	}
 
-	public MaxServiceCurve_DNC_Affine createDelayedInfiniteBurstMSC(double delay) {
+	public MaxServiceCurve_DNC_PwAffine createDelayedInfiniteBurstMSC(double delay) {
 		return createDelayedInfiniteBurstMSC(Num.getFactory(Calculator.getInstance().getNumBackend()).create(delay));
 	}
 
-	public MaxServiceCurve_DNC_Affine createDelayedInfiniteBurstMSC(Num delay) {
-		MaxServiceCurve_DNC_Affine msc_dnc = new MaxServiceCurve_DNC_Affine();
+	public MaxServiceCurve_DNC_PwAffine createDelayedInfiniteBurstMSC(Num delay) {
+		MaxServiceCurve_DNC_PwAffine msc_dnc = new MaxServiceCurve_DNC_PwAffine();
 		makeDelayedInfiniteBurst(msc_dnc, delay);
 		return msc_dnc;
 	}
 
-	public MaxServiceCurve_DNC_Affine createRateLatencyMSC(double rate, double latency) {
+	public MaxServiceCurve_DNC_PwAffine createRateLatencyMSC(double rate, double latency) {
 		return createRateLatencyMSC(Num.getFactory(Calculator.getInstance().getNumBackend()).create(rate), Num.getFactory(Calculator.getInstance().getNumBackend()).create(latency));
 	}
 
-	public MaxServiceCurve_DNC_Affine createRateLatencyMSC(Num rate, Num latency) {
-		MaxServiceCurve_DNC_Affine msc_dnc = new MaxServiceCurve_DNC_Affine();
+	public MaxServiceCurve_DNC_PwAffine createRateLatencyMSC(Num rate, Num latency) {
+		MaxServiceCurve_DNC_PwAffine msc_dnc = new MaxServiceCurve_DNC_PwAffine();
 		makeRateLatency(msc_dnc, rate, latency);
 		return msc_dnc;
 	}
@@ -1240,13 +1231,13 @@ public class Curve_DNC_Affine implements Curve_Affine {
 	// ------------------------------------------------------------------------------
 	// Curve assembly
 	// ------------------------------------------------------------------------------
-	private void makeHorizontal(Curve_DNC_Affine c_dnc, Num y) {
+	private void makeHorizontal(Curve_DNC_PwAffine c_dnc, Num y) {
 		LinearSegment_DNC segment = new LinearSegment_DNC(Num.getFactory(Calculator.getInstance().getNumBackend()).createZero(), y,
 				Num.getFactory(Calculator.getInstance().getNumBackend()).createZero(), false);
 		c_dnc.setSegments(new LinearSegment_DNC[] { segment });
 	}
 
-	private void makeDelayedInfiniteBurst(Curve_DNC_Affine c_dnc, Num delay) {
+	private void makeDelayedInfiniteBurst(Curve_DNC_PwAffine c_dnc, Num delay) {
 		if (delay.ltZero()) {
 			throw new IllegalArgumentException("Delayed infinite burst curve must have delay >= 0.0");
 		}
@@ -1263,7 +1254,7 @@ public class Curve_DNC_Affine implements Curve_Affine {
 		c_dnc.is_delayed_infinite_burst = true;
 	}
 
-	private void makePeakRate(Curve_DNC_Affine c_dnc, Num rate) {
+	private void makePeakRate(Curve_DNC_PwAffine c_dnc, Num rate) {
 		if (rate.equals(Num.getFactory(Calculator.getInstance().getNumBackend()).getPositiveInfinity())) {
 			throw new IllegalArgumentException(
 					"Peak rate with rate infinity equals a delayed infinite burst curve with delay < 0.0");
@@ -1282,7 +1273,7 @@ public class Curve_DNC_Affine implements Curve_Affine {
 		c_dnc.is_token_bucket = true; // with burstiness 0
 	}
 
-	private void makeRateLatency(Curve_DNC_Affine c_dnc, Num rate, Num latency) {
+	private void makeRateLatency(Curve_DNC_PwAffine c_dnc, Num rate, Num latency) {
 		if (rate.equals(Num.getFactory(Calculator.getInstance().getNumBackend()).getPositiveInfinity())) {
 			makeDelayedInfiniteBurst(c_dnc, latency);
 			return;
@@ -1307,7 +1298,7 @@ public class Curve_DNC_Affine implements Curve_Affine {
 		c_dnc.is_rate_latency = true;
 	}
 
-	private void makeTokenBucket(Curve_DNC_Affine c_dnc, Num rate, Num burst) {
+	private void makeTokenBucket(Curve_DNC_PwAffine c_dnc, Num rate, Num burst) {
 		if (rate.equals(Num.getFactory(Calculator.getInstance().getNumBackend()).getPositiveInfinity())
 				|| burst.equals(Num.getFactory(Calculator.getInstance().getNumBackend()).getPositiveInfinity())) {
 			makeDelayedInfiniteBurst(c_dnc, Num.getFactory(Calculator.getInstance().getNumBackend()).createZero());
