+++ src/main/java/de/uni_kl/cs/discodnc/curves/disco/pwaffine/MaxServiceCurve_DNC_PwAffine.java
@@ -2,7 +2,7 @@
  * This file is part of the Disco Deterministic Network Calculator.
  *
  * Copyright (C) 2005 - 2007 Frank A. Zdarsky
- * Copyright (C) 2013 - 2018 Steffen Bondorf
+ * Copyright (C) 2016 Steffen Bondorf
  * Copyright (C) 2017+ The DiscoDNC contributors
  *
  * disco | Distributed Computer Systems Lab
@@ -27,68 +27,78 @@
  *
  */
 
-package de.uni_kl.cs.discodnc.curves.dnc.pwaffine;
+package de.uni_kl.cs.discodnc.curves.disco.pwaffine;
 
 import de.uni_kl.cs.discodnc.Calculator;
-import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
+import de.uni_kl.cs.discodnc.curves.MaxServiceCurve;
 
-public class ArrivalCurve_DNC_PwAffine extends Curve_DNC_PwAffine implements ArrivalCurve {
+public class MaxServiceCurve_DNC_PwAffine extends Curve_DNC_PwAffine implements MaxServiceCurve {
     // --------------------------------------------------------------------------------------------------------------
     // Constructors
     // --------------------------------------------------------------------------------------------------------------
-    public ArrivalCurve_DNC_PwAffine() {
+    protected MaxServiceCurve_DNC_PwAffine() {
         super();
     }
 
-    public ArrivalCurve_DNC_PwAffine(int segment_count) {
+    public MaxServiceCurve_DNC_PwAffine(int segment_count) {
         super(segment_count);
     }
 
-    public ArrivalCurve_DNC_PwAffine(Curve curve) {
-        super(curve);
-        forceThroughOrigin();
-        if (Calculator.getInstance().exec_arrival_curve_checks() && !isWideSenseIncreasing()) {
-        // Too strong requirement: !isConcave()
-            System.out.println(toString());
-            throw new RuntimeException("Arrival curves can only be created from wide-sense increasing functions.");
+    public MaxServiceCurve_DNC_PwAffine(Curve curve) {
+        copy(curve);
+
+        if (Calculator.getInstance().exec_max_service_curve_checks() && !isWideSenseIncreasing()) { // too strong
+            // requirement:
+            // !isAlmostConcave()
+            // ) {
+            throw new RuntimeException(
+                    "Maximum service curves can only be created from wide-sense increasing functions.");
         }
+
+        forceThroughOrigin();
     }
 
-    public ArrivalCurve_DNC_PwAffine(String arrival_curve_str) throws Exception {
-        if (arrival_curve_str == null || arrival_curve_str.isEmpty() || arrival_curve_str.length() < 9) {
-        	// Smallest possible string: {(0,0),0}
+    public MaxServiceCurve_DNC_PwAffine(String max_service_curve_str) throws Exception {
+        if (max_service_curve_str == null || max_service_curve_str.isEmpty() || max_service_curve_str.length() < 9) { // Smallest
+            // possible
+            // string:
+            // {(0,0),0}
             throw new RuntimeException("Invalid string representation of a service curve.");
         }
 
-        initializeCurve(arrival_curve_str);
-        forceThroughOrigin();
+        initializeCurve(max_service_curve_str);
 
-        // Too strong requirement: !isConcave()
-        if (Calculator.getInstance().exec_arrival_curve_checks() && !isWideSenseIncreasing()) { 
-            System.out.println(toString());
-            throw new RuntimeException("Arrival curves can only be created from wide-sense increasing functions.");
+        if (Calculator.getInstance().exec_max_service_curve_checks() && !isWideSenseIncreasing()) { // too strong
+            // requirement:
+            // !isAlmostConcave()
+            // ) {
+            throw new RuntimeException(
+                    "Maximum service curves can only be created from wide-sense increasing functions.");
         }
+
+        forceThroughOrigin();
     }
 
     // --------------------------------------------------------------------------------------------------------------
     // Interface Implementations
     // --------------------------------------------------------------------------------------------------------------
     @Override
-    public ArrivalCurve_DNC_PwAffine copy() {
-        ArrivalCurve_DNC_PwAffine ac_copy = new ArrivalCurve_DNC_PwAffine();
-        ac_copy.copy(this);
-        return ac_copy;
+    public MaxServiceCurve_DNC_PwAffine copy() {
+        MaxServiceCurve_DNC_PwAffine msc_copy = new MaxServiceCurve_DNC_PwAffine();
+        msc_copy.copy(this);
+
+        return msc_copy;
     }
 
     @Override
     public boolean equals(Object obj) {
-        return (obj instanceof ArrivalCurve_DNC_PwAffine) && super.equals(obj);
+        return (obj instanceof MaxServiceCurve_DNC_PwAffine) && super.equals(obj);
     }
 
     @Override
     public int hashCode() {
-        return "AC".hashCode() * super.hashCode();
+        return "MSC".hashCode() * super.hashCode();
     }
 
     /**
@@ -98,6 +108,6 @@ public class ArrivalCurve_DNC_PwAffine extends Curve_DNC_PwAffine implements Arr
      */
     @Override
     public String toString() {
-        return "AC" + super.toString();
+        return "MSC" + super.toString();
     }
 }
