+++ src/main/java/de/uni_kl/cs/discodnc/curves/disco/pwaffine/ServiceCurve_DNC_PwAffine.java
@@ -27,52 +27,44 @@
  *
  */
 
-package de.uni_kl.cs.discodnc.curves.dnc.affine;
+package de.uni_kl.cs.discodnc.curves.disco.pwaffine;
 
 import de.uni_kl.cs.discodnc.Calculator;
-import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
+import de.uni_kl.cs.discodnc.curves.ServiceCurve;
 
-public class ArrivalCurve_DNC_Affine extends Curve_DNC_Affine implements ArrivalCurve {
+public class ServiceCurve_DNC_PwAffine extends Curve_DNC_PwAffine implements ServiceCurve {
     // --------------------------------------------------------------------------------------------------------------
     // Constructors
     // --------------------------------------------------------------------------------------------------------------
-    public ArrivalCurve_DNC_Affine() {
+    public ServiceCurve_DNC_PwAffine() {
         super();
     }
 
-    public ArrivalCurve_DNC_Affine(int segment_count) {
+    public ServiceCurve_DNC_PwAffine(int segment_count) {
         super(segment_count);
     }
 
-    public ArrivalCurve_DNC_Affine(Curve curve) {
-        super(curve);
-        forceThroughOrigin();
+    public ServiceCurve_DNC_PwAffine(Curve curve) {
+        copy(curve);
 
-        if (Calculator.getInstance().exec_arrival_curve_checks() && !isWideSenseIncreasing()) { // too strong
-            // requirement:
-            // !isConcave()
-            System.out.println(toString());
-            throw new RuntimeException("Arrival curves can only be created from wide-sense increasing functions.");
+        // Too strong requirement: !isConvex()
+        if (Calculator.getInstance().exec_service_curve_checks() && !isWideSenseIncreasing()) {
+            throw new RuntimeException("Service curves can only be created from wide-sense increasing functions.");
         }
     }
 
-    public ArrivalCurve_DNC_Affine(String arrival_curve_str) throws Exception {
-        if (arrival_curve_str == null || arrival_curve_str.isEmpty() || arrival_curve_str.length() < 9) { // Smallest
-            // possible
-            // string:
-            // {(0,0),0}
+    public ServiceCurve_DNC_PwAffine(String service_curve_str) throws Exception {
+        if (service_curve_str == null || service_curve_str.isEmpty() || service_curve_str.length() < 9) {
+        	// Smallest possible string: {(0,0),0}
             throw new RuntimeException("Invalid string representation of a service curve.");
         }
 
-        initializeCurve(arrival_curve_str);
-        forceThroughOrigin();
+        initializeCurve(service_curve_str);
 
-        if (Calculator.getInstance().exec_arrival_curve_checks() && !isWideSenseIncreasing()) { // too strong
-            // requirement:
-            // !isConcave()
-            System.out.println(toString());
-            throw new RuntimeException("Arrival curves can only be created from wide-sense increasing functions.");
+        // Too strong requirement: !isConvex()
+        if (Calculator.getInstance().exec_service_curve_checks() && !isWideSenseIncreasing()) {
+            throw new RuntimeException("Service curves can only be created from wide-sense increasing functions.");
         }
     }
 
@@ -80,20 +72,20 @@ public class ArrivalCurve_DNC_Affine extends Curve_DNC_Affine implements Arrival
     // Interface Implementations
     // --------------------------------------------------------------------------------------------------------------
     @Override
-    public ArrivalCurve_DNC_Affine copy() {
-        ArrivalCurve_DNC_Affine ac_copy = new ArrivalCurve_DNC_Affine();
-        ac_copy.copy(this);
-        return ac_copy;
+    public ServiceCurve_DNC_PwAffine copy() {
+        ServiceCurve_DNC_PwAffine sc_copy = new ServiceCurve_DNC_PwAffine();
+        sc_copy.copy(this);
+        return sc_copy;
     }
 
     @Override
     public boolean equals(Object obj) {
-        return (obj instanceof ArrivalCurve_DNC_Affine) && super.equals(obj);
+        return (obj instanceof ServiceCurve_DNC_PwAffine) && super.equals(obj);
     }
 
     @Override
     public int hashCode() {
-        return "AC".hashCode() * super.hashCode();
+        return "SC".hashCode() * super.hashCode();
     }
 
     /**
@@ -103,6 +95,6 @@ public class ArrivalCurve_DNC_Affine extends Curve_DNC_Affine implements Arrival
      */
     @Override
     public String toString() {
-        return "AC" + super.toString();
+        return "SC" + super.toString();
     }
 }
